Data structure revamp. Start of collaborative editing.

+++ src/components/Item.tsx
@@ -1,7 +1,9 @@
 import React from "react";
 import classNames from "classnames";
+import { useAuthState } from "react-firebase-hooks/auth";
 
-import { firestore } from "../firebase";
+import { auth, firestore } from "../firebase";
+import { ShoppingList } from "../interfaces/ShoppingList";
 import { ShoppingListItem } from "../interfaces/ShoppingListItem";
 
 import Alert from "./Alert";
@@ -9,21 +11,43 @@ import Stack from "./Stack";
 import Text from "./Text";
 
 interface Props {
+  shoppingList: ShoppingList;
   item: ShoppingListItem;
 }
 
-const Item: React.FC<Props> = ({ item }) => {
+const Item: React.FC<Props> = ({ shoppingList, item }) => {
+  const [user] = useAuthState(auth);
   const [checked, setChecked] = React.useState(false);
   const [error, setError] = React.useState("");
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
     if (e.target.checked) {
+      if (!user) {
+        setError("Invalid user");
+
+        return;
+      }
+
       setChecked(true);
 
       firestore
+        .collection("shoppingLists")
+        .doc(shoppingList.id)
         .collection("items")
         .doc(item.id)
         .delete()
+        .then(() => {
+          firestore
+            .collection("shoppingLists")
+            .doc(shoppingList.id)
+            .update({
+              updatedAt: new Date(),
+              updatedBy: user.uid,
+            })
+            .catch((err: string) => {
+              setError(err);
+            });
+        })
         .catch((err: string) => {
           setError(err);
         });
